refactor(desktop): split context panel into strips/parts/slot-row
