+++ apps/desktop/src/features/context/components/ContextPanel/index.tsx
@@ -1,48 +1,10 @@
-import { useEffect, useState, useCallback, useMemo } from 'react';
-import {
-  AlertTriangle,
-  ArrowUpRight,
-  PanelRightClose,
-  PanelRightOpen,
-  History,
-  RotateCcw,
-  RotateCw,
-  ChevronRight,
-  ChevronDown,
-  Clock,
-  FileEdit,
-  GitMerge,
-  GitPullRequest,
-  GitPullRequestArrow,
-  Loader2,
-  MessageSquare,
-  RefreshCw,
-  Target,
-  CheckCheck,
-  HelpCircle,
-  Activity,
-  ClipboardList,
-  BookOpen,
-  SquareTerminal,
-  Upload,
-  XCircle,
-  type LucideIcon,
-} from 'lucide-react';
-import { Divider, ScrollFade, Textarea, Dialog, Markdown, cn } from '@goodboy/ui';
-import { SLOT_KEYS, SLOT_LABELS, type SlotKey } from '@goodboy/core';
-import type {
-  ContextSlot,
-  ContextSlotHistoryEntry,
-  PlanStatus,
-  PrCheckRun,
-  Session,
-  SessionId,
-  TelemetryRecord,
-} from '@goodboy/types';
+import { useEffect, useState, useMemo } from 'react';
+import { PanelRightClose, PanelRightOpen } from 'lucide-react';
+import { Divider, ScrollFade, cn } from '@goodboy/ui';
+import { SLOT_KEYS } from '@goodboy/core';
+import type { ContextSlot, Session, SessionId, TelemetryRecord } from '@goodboy/types';
 import { OpenQuestionsStrip } from '../OpenQuestionsStrip';
-import { PullRequestChip } from '../../../../features/github/components/PullRequestChip';
 import { worktreeStatus } from '../../../../features/worktree/worktree';
-import { useRemoteHostKind } from '../../../../features/worktree/useRemoteHostKind';
 import { TerminalDock } from '../../../../features/terminal/components/TerminalDock';
 import {
   EMPTY_ARRAY,
@@ -51,9 +13,13 @@ import {
   useSessionLoading,
   useSessionPlans,
   useSessionSlots,
-  useSlotHistory,
   useSummarizerStatus,
 } from '../../../../store';
+import { ICON_BTN, normalizeFilesSlot, type PanelTab } from './lib';
+import { TabStrip } from './parts/TabStrip';
+import { SummarizerBadge } from './parts/SummarizerBadge';
+import { ChangesStrip } from './strips/ChangesStrip';
+import { SlotRow } from './SlotRow';
 
 interface ContextPanelProps {
   session: Session;
@@ -62,13 +28,6 @@ interface ContextPanelProps {
   onExpand?: () => void;
   isActive?: boolean;
 }
-
-type SummarizerStatusKind = 'idle' | 'running' | 'error';
-
-const ICON_BTN =
-  'rounded-sm p-0.5 text-muted-foreground hover:bg-muted hover:text-foreground' as const;
-
-type PanelTab = 'context' | 'terminal';
 
 export function ContextPanel({
   session,
@@ -303,950 +262,4 @@ export function ContextPanel({
       </div>
     </>
   );
-}
-
-interface TabStripProps {
-  readonly tab: PanelTab;
-  readonly onPick: (next: PanelTab) => void;
-  readonly summarizerRunning: boolean;
-  readonly isTerminalOpen: boolean;
-}
-
-function TabStrip({ tab, onPick, summarizerRunning, isTerminalOpen }: TabStripProps) {
-  return (
-    <div role="tablist" aria-label="context panel tabs" className="flex items-center gap-0.5">
-      <TabButton
-        active={tab === 'context'}
-        onClick={() => onPick('context')}
-        icon={<BookOpen size={11} aria-hidden />}
-        label="Context"
-        accentDot={summarizerRunning ? 'bg-info' : null}
-      />
-      <TabButton
-        active={tab === 'terminal'}
-        onClick={() => onPick('terminal')}
-        icon={<SquareTerminal size={11} aria-hidden />}
-        label="Terminal"
-        accentDot={isTerminalOpen ? 'bg-info' : null}
-      />
-    </div>
-  );
-}
-
-interface TabButtonProps {
-  readonly active: boolean;
-  readonly onClick: () => void;
-  readonly icon: React.ReactNode;
-  readonly label: string;
-  readonly badge?: number | null;
-  readonly accentDot?: string | null;
-}
-
-function TabButton({ active, onClick, icon, label, badge, accentDot }: TabButtonProps) {
-  return (
-    <button
-      type="button"
-      role="tab"
-      aria-selected={active}
-      aria-label={label}
-      title={label}
-      onClick={onClick}
-      className={cn(
-        'inline-flex items-center gap-1 rounded-md px-2 py-1 text-2xs font-semibold uppercase leading-none tracking-[0.06em] transition-colors',
-        active
-          ? 'bg-background text-foreground shadow-sm'
-          : 'text-muted-foreground hover:bg-foreground/5 hover:text-foreground',
-      )}
-    >
-      {icon}
-      {active ? <span>{label}</span> : null}
-      {badge !== null && badge !== undefined ? (
-        <span className="ml-0.5 inline-flex min-w-[1rem] items-center justify-center rounded-full bg-muted px-1 text-[9px] font-medium tracking-normal text-muted-foreground">
-          {badge}
-        </span>
-      ) : null}
-      {accentDot ? (
-        <span aria-hidden className={cn('ml-0.5 size-1.5 rounded-full', accentDot)} />
-      ) : null}
-    </button>
-  );
-}
-
-interface FilesTouchedShape {
-  readonly paths: ReadonlyArray<string>;
-  readonly count: number;
-  readonly additions: number;
-  readonly deletions: number;
-}
-
-const PLAN_STATUS_STYLE: Record<PlanStatus, string> = {
-  active: 'bg-warning/10 text-warning',
-  consumed: 'bg-info/10 text-info',
-  superseded: 'bg-muted text-muted-foreground',
-  discarded: 'bg-muted/60 text-muted-foreground/70 line-through',
-};
-
-const PLAN_STATUS_LABEL: Record<PlanStatus, string> = {
-  active: 'Active',
-  consumed: 'Consumed',
-  superseded: 'Superseded',
-  discarded: 'Discarded',
-};
-
-interface ChangesStripProps {
-  sessionId: SessionId;
-  workingDir: string | null;
-  filesTouched: FilesTouchedShape;
-  plans: ReadonlyArray<{ id: string; status: PlanStatus; title: string }>;
-  plansLoading: boolean;
-  hasActivePlan: boolean;
-}
-
-function ChangesStrip({
-  sessionId,
-  workingDir,
-  filesTouched,
-  plans,
-  plansLoading,
-  hasActivePlan,
-}: ChangesStripProps) {
-  const count = filesTouched.count;
-
-  const openPlanStudio = () =>
-    window.dispatchEvent(new CustomEvent('goodboy:open-plan-studio', { detail: { sessionId } }));
-
-  const openDiffViewer = () =>
-    window.dispatchEvent(
-      new CustomEvent('goodboy:open-diff-viewer', { detail: { sessionId, workingDir } }),
-    );
-
-  return (
-    <div className="flex shrink-0 flex-col gap-2 border-t border-border-soft/50 pt-2.5">
-      {plans.length === 0 && plansLoading ? (
-        <PlansSkeleton />
-      ) : (
-        <button
-          type="button"
-          onClick={openPlanStudio}
-          title="Open Plan Studio"
-          aria-label="Open Plan Studio"
-          className={cn(
-            'flex w-full items-center justify-between gap-2 rounded-lg px-3 py-2 text-xs ring-1 transition-colors hover:bg-foreground/5',
-            plans.length > 0
-              ? 'ring-border-soft'
-              : 'text-muted-foreground/70 ring-border-soft/40 hover:text-foreground',
-          )}
-        >
-          <span className="inline-flex min-w-0 items-center gap-1.5">
-            <ClipboardList size={12} aria-hidden />
-            {plans.length > 0 ? (
-              <span className="font-medium">
-                {plans.length} plan{plans.length === 1 ? '' : 's'}
-              </span>
-            ) : (
-              <span>No plans yet. Spawn a Plan agent and ask it to map the work.</span>
-            )}
-            {hasActivePlan ? (
-              <span
-                className={cn(
-                  'inline-flex shrink-0 items-center justify-center rounded-full px-1.5 py-0.5 text-[9px] uppercase tracking-wide',
-                  PLAN_STATUS_STYLE['active'],
-                )}
-              >
-                {PLAN_STATUS_LABEL['active']}
-              </span>
-            ) : null}
-          </span>
-          <ArrowUpRight size={12} aria-hidden className="shrink-0 opacity-70" />
-        </button>
-      )}
-      <GithubStrip sessionId={sessionId} />
-      <GitlabMrStrip sessionId={sessionId} />
-      <PendingResolutionsStrip sessionId={sessionId} />
-      {count > 0 ? (
-        <button
-          type="button"
-          onClick={openDiffViewer}
-          disabled={!workingDir}
-          title={workingDir ? 'open the diff viewer' : 'no worktree for this session'}
-          className="flex w-full items-center justify-between gap-2 rounded-lg bg-info/5 px-3 py-2 text-xs text-info ring-1 ring-info/20 transition-colors hover:bg-info/10 disabled:cursor-default disabled:opacity-60"
-        >
-          <span className="inline-flex min-w-0 items-center gap-1.5">
-            <FileEdit size={12} aria-hidden />
-            <span className="truncate font-medium">
-              {count} file{count === 1 ? '' : 's'} touched
-            </span>
-          </span>
-          <span className="inline-flex shrink-0 items-center gap-2">
-            {filesTouched.additions > 0 || filesTouched.deletions > 0 ? (
-              <span className="inline-flex items-center gap-1 font-mono text-[10px] tabular-nums">
-                {filesTouched.additions > 0 ? (
-                  <span className="text-success">+{filesTouched.additions}</span>
-                ) : null}
-                {filesTouched.deletions > 0 ? (
-                  <span className="text-danger">−{filesTouched.deletions}</span>
-                ) : null}
-              </span>
-            ) : null}
-            <ArrowUpRight size={12} aria-hidden className="opacity-70" />
-          </span>
-        </button>
-      ) : (
-        <div className="flex w-full items-center gap-1.5 rounded-lg px-3 py-2 text-xs text-muted-foreground/60 ring-1 ring-border-soft/40">
-          <FileEdit size={12} aria-hidden />
-          <span className="font-medium">working tree clean</span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function PendingResolutionsStrip({ sessionId }: { sessionId: SessionId }) {
-  const pending = useAppStore((s) => s.sessionPendingResolutions[sessionId] ?? EMPTY_ARRAY);
-  const loadPendingResolutions = useAppStore((s) => s.loadPendingResolutions);
-  const pushAllResolutions = useAppStore((s) => s.pushAllResolutions);
-  const [busy, setBusy] = useState(false);
-
-  useEffect(() => {
-    void loadPendingResolutions(sessionId);
-  }, [sessionId, loadPendingResolutions]);
-
-  const count = pending.length;
-  if (count === 0) return null;
-
-  const onPush = async () => {
-    if (busy) return;
-    setBusy(true);
-    try {
-      await pushAllResolutions(sessionId);
-    } finally {
-      setBusy(false);
-    }
-  };
-
-  return (
-    <button
-      type="button"
-      onClick={() => void onPush()}
-      disabled={busy}
-      title="push the branch once, then reply and resolve every queued review comment"
-      className="flex w-full items-center justify-between gap-2 rounded-lg bg-accent/5 px-3 py-2 text-xs text-accent ring-1 ring-accent/20 transition-colors hover:bg-accent/10 disabled:cursor-default disabled:opacity-60"
-    >
-      <span className="inline-flex min-w-0 items-center gap-1.5">
-        <Upload size={12} aria-hidden />
-        <span className="truncate font-medium">
-          Push &amp; resolve {count} comment{count === 1 ? '' : 's'}
-        </span>
-      </span>
-      {busy ? (
-        <Loader2 size={12} aria-hidden className="shrink-0 animate-spin" />
-      ) : (
-        <ArrowUpRight size={12} aria-hidden className="shrink-0 opacity-70" />
-      )}
-    </button>
-  );
-}
-
-function GithubStrip({ sessionId }: { sessionId: SessionId }) {
-  const workspaceId = useAppStore(
-    (s) => s.sessions.find((x) => x.id === sessionId)?.workspaceId ?? null,
-  );
-  const remoteKind = useRemoteHostKind(workspaceId);
-  const github = useAppStore((s) => s.sessionGithub[sessionId]);
-  const refreshSessionPr = useAppStore((s) => s.refreshSessionPr);
-  const pr = github?.pr ?? null;
-  const loading = github?.loading ?? false;
-  const error = github?.error ?? null;
-  const openStudio = () =>
-    window.dispatchEvent(new CustomEvent('goodboy:open-github-session', { detail: { sessionId } }));
-
-  const detail = github?.detail ?? null;
-  const unresolvedComments = (detail?.comments ?? []).filter(
-    (c) => c.source !== 'review' || c.resolved === false,
-  ).length;
-  const ciState = computeCiState(detail?.checks ?? []);
-
-  if (remoteKind === 'gitlab') {
-    return null;
-  }
-
-  return (
-    <div className="flex flex-col gap-1">
-      <div className="flex items-center gap-1">
-        <button
-          type="button"
-          onClick={openStudio}
-          title="open in github studio"
-          className={cn(
-            'flex min-w-0 flex-1 items-center justify-between gap-2 rounded-lg px-3 py-2 text-xs ring-1 transition-colors hover:bg-foreground/5',
-            pr
-              ? 'ring-border-soft'
-              : 'text-muted-foreground/70 ring-border-soft/40 hover:text-foreground',
-          )}
-        >
-          {pr ? (
-            <span className="inline-flex min-w-0 items-center gap-2">
-              <PullRequestChip state={pr.state} variant="badge" number={pr.number} iconSize={11} />
-              {ciState !== 'none' ? <CiBadge state={ciState} /> : null}
-              {unresolvedComments > 0 ? (
-                <span className="inline-flex items-center gap-1 text-2xs text-muted-foreground">
-                  <MessageSquare size={11} aria-hidden />
-                  <span className="tabular-nums">{unresolvedComments}</span>
-                </span>
-              ) : null}
-            </span>
-          ) : (
-            <span className="inline-flex items-center gap-1.5">
-              <GitPullRequest size={12} aria-hidden />
-              <span>No PR yet</span>
-            </span>
-          )}
-          <ArrowUpRight size={12} aria-hidden className="shrink-0 opacity-70" />
-        </button>
-        <button
-          type="button"
-          onClick={() => void refreshSessionPr(sessionId, { force: true })}
-          disabled={loading}
-          title={error ? `refresh failed: ${error}` : 'refresh PR status'}
-          aria-label="refresh PR status"
-          className="flex size-7 shrink-0 items-center justify-center rounded-md text-muted-foreground ring-1 ring-border-soft/40 transition-colors hover:bg-foreground/5 hover:text-foreground disabled:opacity-50"
-        >
-          <RefreshCw size={12} aria-hidden className={loading ? 'animate-spin' : undefined} />
-        </button>
-      </div>
-      {error ? (
-        <span className="px-1 text-2xs text-danger" title={error}>
-          {error}
-        </span>
-      ) : null}
-    </div>
-  );
-}
-
-function GitlabMrStrip({ sessionId }: { sessionId: SessionId }) {
-  const workspaceId = useAppStore(
-    (s) => s.sessions.find((x) => x.id === sessionId)?.workspaceId ?? null,
-  );
-  const remoteKind = useRemoteHostKind(workspaceId);
-  const mrState = useAppStore((s) => s.sessionGitlabMr[sessionId]);
-  const refreshSessionMr = useAppStore((s) => s.refreshSessionMr);
-  const mr = mrState?.mr ?? null;
-  const loading = mrState?.loading ?? false;
-  const error = mrState?.error ?? null;
-  const openPane = () =>
-    window.dispatchEvent(new CustomEvent('goodboy:open-gitlab-mr', { detail: { sessionId } }));
-
-  if (remoteKind !== 'gitlab') {
-    return null;
-  }
-
-  return (
-    <div className="flex flex-col gap-1">
-      <div className="flex items-center gap-1">
-        <button
-          type="button"
-          onClick={openPane}
-          title="open merge request"
-          className={cn(
-            'flex min-w-0 flex-1 items-center justify-between gap-2 rounded-lg px-3 py-2 text-xs ring-1 transition-colors hover:bg-foreground/5',
-            mr
-              ? 'ring-border-soft'
-              : 'text-muted-foreground/70 ring-border-soft/40 hover:text-foreground',
-          )}
-        >
-          {mr ? (
-            <span className="inline-flex min-w-0 items-center gap-2">
-              <span className="inline-flex items-center gap-1 font-medium">
-                <GitMerge size={11} aria-hidden />!{mr.iid}
-              </span>
-              <span className="text-2xs text-muted-foreground">{mr.state}</span>
-              {mr.draft ? (
-                <span className="rounded bg-warning/15 px-1 py-px text-[9px] font-medium uppercase tracking-wide text-warning">
-                  draft
-                </span>
-              ) : null}
-            </span>
-          ) : (
-            <span className="inline-flex items-center gap-1.5">
-              <GitMerge size={12} aria-hidden />
-              <span>No MR yet</span>
-            </span>
-          )}
-          <ArrowUpRight size={12} aria-hidden className="shrink-0 opacity-70" />
-        </button>
-        <button
-          type="button"
-          onClick={() => void refreshSessionMr(sessionId, { force: true })}
-          disabled={loading}
-          title={error ? `refresh failed: ${error}` : 'refresh MR status'}
-          aria-label="refresh MR status"
-          className="flex size-7 shrink-0 items-center justify-center rounded-md text-muted-foreground ring-1 ring-border-soft/40 transition-colors hover:bg-foreground/5 hover:text-foreground disabled:opacity-50"
-        >
-          <RefreshCw size={12} aria-hidden className={loading ? 'animate-spin' : undefined} />
-        </button>
-      </div>
-      {error ? (
-        <span className="px-1 text-2xs text-danger" title={error}>
-          {error}
-        </span>
-      ) : null}
-    </div>
-  );
-}
-
-type CiState = 'success' | 'failure' | 'pending' | 'none';
-
-function computeCiState(checks: ReadonlyArray<PrCheckRun>): CiState {
-  if (checks.length === 0) return 'none';
-  if (
-    checks.some(
-      (c) =>
-        c.conclusion === 'failure' || c.conclusion === 'cancelled' || c.conclusion === 'timed_out',
-    )
-  ) {
-    return 'failure';
-  }
-  if (checks.some((c) => c.conclusion === 'pending')) return 'pending';
-  if (checks.some((c) => c.conclusion === 'success')) return 'success';
-  return 'none';
-}
-
-function CiBadge({ state }: { state: CiState }) {
-  const map: Record<CiState, { icon: LucideIcon; className: string; label: string }> = {
-    success: { icon: GitPullRequestArrow, className: 'text-success', label: 'ci ✓' },
-    failure: { icon: XCircle, className: 'text-danger', label: 'ci ✗' },
-    pending: { icon: Clock, className: 'text-warning', label: 'ci …' },
-    none: { icon: Clock, className: 'text-muted-foreground/40', label: 'no ci' },
-  };
-  const entry = map[state];
-  const Icon = entry.icon;
-  return (
-    <span className={cn('inline-flex items-center gap-0.5', entry.className)}>
-      <Icon size={11} aria-hidden />
-      <span className="font-medium">{entry.label}</span>
-    </span>
-  );
-}
-
-function SlotRowSkeleton({ slotKey }: { slotKey: SlotKey }) {
-  return (
-    <li role="status" aria-label={`loading ${slotKey}`} className="flex flex-col gap-2">
-      <div className="h-2.5 w-20 rounded bg-muted/50" />
-      <div className="h-3 w-full rounded bg-muted/50" />
-      <div className="h-3 w-3/4 rounded bg-muted/50" />
-    </li>
-  );
-}
-
-function PlansSkeleton() {
-  return (
-    <div role="status" aria-label="loading plans" className="flex flex-col gap-2 pb-2">
-      <div className="h-2.5 w-16 rounded bg-muted/50" />
-      <div className="h-3 w-full rounded bg-muted/50" />
-      <div className="h-3 w-2/3 rounded bg-muted/50" />
-    </div>
-  );
-}
-
-interface SlotMeta {
-  readonly icon: LucideIcon;
-  readonly iconClass: string;
-  readonly iconChipBg: string;
-  readonly description: string;
-  readonly emphasis?: boolean;
-  readonly accentRingWhenNonEmpty?: string;
-  readonly emptyLabel: string;
-  readonly emptyCta: string;
-  readonly collapsible?: boolean;
-  readonly defaultCollapsed?: boolean;
-  readonly singleLine?: boolean;
-  readonly readOnly?: boolean;
-}
-
-const MARKDOWN_SLOTS: ReadonlySet<SlotKey> = new Set<SlotKey>([
-  'open_questions',
-  'decisions',
-  'last_output_summary',
-]);
-
-const SLOT_META: Record<Exclude<SlotKey, 'files_touched'>, SlotMeta> = {
-  goal: {
-    icon: Target,
-    iconClass: 'text-primary',
-    iconChipBg: 'bg-primary/10 ring-primary/20',
-    description: 'What this session is meant to achieve',
-    emphasis: true,
-    singleLine: true,
-    emptyLabel: 'No goal yet',
-    emptyCta: 'Add the session goal',
-  },
-  open_questions: {
-    icon: HelpCircle,
-    iconClass: 'text-warning',
-    iconChipBg: 'bg-warning/10 ring-warning/20',
-    description: 'Things the agent still needs clarified',
-    accentRingWhenNonEmpty: 'ring-warning/60',
-    emptyLabel: 'No open questions',
-    emptyCta: '',
-    readOnly: true,
-  },
-  decisions: {
-    icon: CheckCheck,
-    iconClass: 'text-success',
-    iconChipBg: 'bg-success/10 ring-success/20',
-    description: 'Choices already locked in for this session',
-    collapsible: true,
-    defaultCollapsed: true,
-    emptyLabel: 'No decisions yet',
-    emptyCta: 'Log a decision',
-  },
-  last_output_summary: {
-    icon: Activity,
-    iconClass: 'text-info',
-    iconChipBg: 'bg-info/10 ring-info/20',
-    description: "Summary of the agent's most recent reply",
-    collapsible: true,
-    defaultCollapsed: true,
-    emptyLabel: 'No output yet',
-    emptyCta: 'Write a manual summary',
-  },
-};
-
-function countMarkdownItems(value: string): number {
-  return value.split('\n').filter((l) => /^\s*([-*+]|\d+\.)\s+/.test(l)).length;
-}
-
-function firstMeaningfulLine(value: string): string {
-  for (const raw of value.split('\n')) {
-    const t = raw
-      .trim()
-      .replace(/^[-*+]\s+/, '')
-      .replace(/^\d+\.\s+/, '')
-      .replace(/^#+\s+/, '')
-      .replace(/\*\*/g, '')
-      .replace(/`/g, '');
-    if (t) return t;
-  }
-  return '';
-}
-
-function normalizeFilesSlot(slot: ContextSlot, workingDir: string | null): ContextSlot {
-  if (!workingDir || slot.value.length === 0) return slot;
-  const root = workingDir.endsWith('/') ? workingDir : `${workingDir}/`;
-  const normalized = slot.value
-    .split('\n')
-    .map((p) => (p.startsWith(root) ? p.slice(root.length) : p))
-    .join('\n');
-  return normalized === slot.value ? slot : { ...slot, value: normalized };
-}
-
-interface SlotRowProps {
-  sessionId: SessionId;
-  slotKey: SlotKey;
-  slot: ContextSlot | undefined;
-  loading?: boolean;
-  isSummarizing?: boolean;
-  onCommit: (value: string) => void;
-}
-
-function SlotRow({
-  sessionId,
-  slotKey,
-  slot,
-  loading = false,
-  isSummarizing = false,
-  onCommit,
-}: SlotRowProps) {
-  const value = slot?.value ?? '';
-  const meta = slotKey === 'files_touched' ? null : SLOT_META[slotKey];
-
-  const [editing, setEditing] = useState(false);
-  const [draft, setDraft] = useState(value);
-  const [historyOpen, setHistoryOpen] = useState(false);
-  const [collapsed, setCollapsed] = useState(meta?.defaultCollapsed ?? false);
-
-  const loadSlotHistory = useAppStore((s) => s.loadSlotHistory);
-  const history = useSlotHistory(sessionId, slotKey);
-
-  useEffect(() => {
-    if (!editing) setDraft(value);
-  }, [value, editing]);
-
-  const openHistory = useCallback(() => {
-    void loadSlotHistory(sessionId, slotKey);
-    setHistoryOpen(true);
-  }, [loadSlotHistory, sessionId, slotKey]);
-
-  const restore = useCallback(
-    (entry: ContextSlotHistoryEntry) => {
-      onCommit(entry.value);
-      setHistoryOpen(false);
-    },
-    [onCommit],
-  );
-
-  if (slot === undefined && loading) return <SlotRowSkeleton slotKey={slotKey} />;
-
-  const Icon = meta?.icon;
-  const hasValue = value.length > 0;
-  const renderAsMarkdown = MARKDOWN_SLOTS.has(slotKey);
-  const collapsible = meta?.collapsible ?? false;
-  const singleLine = meta?.singleLine ?? false;
-  const itemCount = collapsible && hasValue ? countMarkdownItems(value) : 0;
-  const preview = collapsible && hasValue ? firstMeaningfulLine(value) : '';
-  const CollapseIcon = collapsed ? ChevronRight : ChevronDown;
-
-  const commit = () => {
-    setEditing(false);
-    if (draft !== value) onCommit(draft);
-  };
-
-  const itemCountLabel =
-    hasValue && itemCount > 0 ? `${itemCount} item${itemCount === 1 ? '' : 's'}` : null;
-
-  const headerToggle = collapsible ? (
-    <button
-      type="button"
-      onClick={() => setCollapsed((v) => !v)}
-      title={collapsed ? 'Expand' : 'Collapse'}
-      aria-expanded={!collapsed}
-      className="rounded p-0.5 text-muted-foreground/50 transition-colors hover:bg-foreground/10 hover:text-foreground"
-    >
-      <CollapseIcon size={11} aria-hidden />
-    </button>
-  ) : null;
-
-  const inactive = !hasValue;
-  const ringClass = inactive
-    ? 'ring-border-soft/30'
-    : (meta?.accentRingWhenNonEmpty ?? 'ring-border-soft');
-
-  return (
-    <li
-      className={cn(
-        'group relative flex flex-none flex-col gap-2 rounded-lg p-3 ring-1 transition-colors',
-        inactive ? 'bg-transparent' : 'bg-muted/40',
-        ringClass,
-      )}
-    >
-      <div className="flex items-center gap-2">
-        {Icon ? (
-          <span
-            aria-hidden
-            className={cn(
-              'flex size-5 shrink-0 items-center justify-center rounded-md ring-1',
-              meta?.iconChipBg,
-            )}
-          >
-            <Icon size={11} className={meta?.iconClass} aria-hidden />
-          </span>
-        ) : null}
-        <div className="flex min-w-0 flex-1 flex-col">
-          <span className="flex items-baseline gap-1.5 text-2xs font-semibold uppercase tracking-[0.08em] text-foreground">
-            <span>{SLOT_LABELS[slotKey]}</span>
-            {itemCountLabel ? (
-              <span className="text-[10px] font-normal normal-case tracking-normal text-muted-foreground/60">
-                · {itemCountLabel}
-              </span>
-            ) : null}
-          </span>
-          {meta?.description ? (
-            <span className="text-[10px] leading-tight text-muted-foreground/60">
-              {meta.description}
-            </span>
-          ) : null}
-        </div>
-        {history.length > 0 ? (
-          <button
-            type="button"
-            onClick={openHistory}
-            title="View history"
-            aria-label={`view history for ${SLOT_LABELS[slotKey]}`}
-            className="shrink-0 rounded p-0.5 text-muted-foreground/50 transition-colors hover:bg-foreground/10 hover:text-foreground"
-          >
-            <History size={11} aria-hidden />
-          </button>
-        ) : null}
-        {headerToggle}
-      </div>
-
-      {editing ? (
-        <Textarea
-          autoFocus
-          value={draft}
-          onChange={(e) => setDraft(e.target.value)}
-          onBlur={commit}
-          onKeyDown={(e) => {
-            if (e.key === 'Escape') {
-              e.preventDefault();
-              setDraft(value);
-              setEditing(false);
-            }
-            if (e.key === 'Enter' && (e.metaKey || e.ctrlKey)) {
-              e.preventDefault();
-              commit();
-            }
-          }}
-          className="font-mono text-xs"
-          autoGrow
-          maxRows={16}
-        />
-      ) : !hasValue ? (
-        meta?.readOnly ? (
-          <span className="text-xs text-muted-foreground/50">{meta.emptyLabel}</span>
-        ) : (
-          <button
-            type="button"
-            onClick={() => {
-              if (isSummarizing) return;
-              setEditing(true);
-            }}
-            disabled={isSummarizing}
-            className={cn(
-              'flex flex-col items-start gap-0.5 rounded text-left text-xs text-muted-foreground/60 transition-colors',
-              isSummarizing ? 'cursor-default' : 'hover:text-foreground',
-            )}
-          >
-            <span>{meta?.emptyLabel ?? 'Empty'}</span>
-            <span className="text-[10px] text-muted-foreground/40 underline-offset-2 group-hover:underline">
-              {meta?.emptyCta ?? 'Click to edit'}
-            </span>
-          </button>
-        )
-      ) : singleLine ? (
-        <button
-          type="button"
-          onClick={() => {
-            if (isSummarizing) return;
-            setEditing(true);
-          }}
-          disabled={isSummarizing}
-          title={value}
-          className={cn(
-            'rounded text-left text-sm font-medium leading-snug text-foreground transition-colors',
-            isSummarizing ? 'cursor-default' : 'cursor-text hover:bg-foreground/5',
-          )}
-        >
-          {value}
-        </button>
-      ) : collapsible && collapsed ? (
-        <button
-          type="button"
-          onClick={() => setCollapsed(false)}
-          title="Expand"
-          className="flex w-full cursor-pointer items-center gap-1.5 truncate rounded text-left text-xs text-muted-foreground transition-colors hover:text-foreground"
-        >
-          <span className="truncate">{preview}</span>
-        </button>
-      ) : renderAsMarkdown ? (
-        <div
-          role={isSummarizing ? undefined : 'button'}
-          tabIndex={isSummarizing ? -1 : 0}
-          onClick={() => {
-            if (isSummarizing) return;
-            setEditing(true);
-          }}
-          onKeyDown={(e) => {
-            if (isSummarizing) return;
-            if (e.key === 'Enter' || e.key === ' ') {
-              e.preventDefault();
-              setEditing(true);
-            }
-          }}
-          className={cn(
-            'rounded text-left leading-relaxed transition-colors [overflow-wrap:anywhere] focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-primary/15 [&_code]:break-all [&_pre]:whitespace-pre-wrap [&_pre]:break-all',
-            isSummarizing ? 'cursor-default' : 'cursor-text',
-          )}
-        >
-          <Markdown text={value} className="text-[13px] text-foreground" />
-        </div>
-      ) : (
-        <button
-          type="button"
-          onClick={() => {
-            if (isSummarizing) return;
-            setEditing(true);
-          }}
-          disabled={isSummarizing}
-          className={cn(
-            'whitespace-pre-wrap break-words rounded text-left leading-relaxed transition-colors',
-            isSummarizing ? 'cursor-default' : 'cursor-text hover:bg-foreground/5',
-            meta?.emphasis ? 'text-sm font-medium' : 'text-xs',
-          )}
-        >
-          {value}
-        </button>
-      )}
-
-      <SlotHistoryDialog
-        label={SLOT_LABELS[slotKey]}
-        renderAsMarkdown={renderAsMarkdown}
-        open={historyOpen}
-        entries={history}
-        onRestore={restore}
-        onClose={() => setHistoryOpen(false)}
-      />
-    </li>
-  );
-}
-
-interface SlotHistoryDialogProps {
-  label: string;
-  renderAsMarkdown: boolean;
-  open: boolean;
-  entries: ReadonlyArray<ContextSlotHistoryEntry>;
-  onRestore: (entry: ContextSlotHistoryEntry) => void;
-  onClose: () => void;
-}
-
-function SlotHistoryDialog({
-  label,
-  renderAsMarkdown,
-  open,
-  entries,
-  onRestore,
-  onClose,
-}: SlotHistoryDialogProps) {
-  return (
-    <Dialog open={open} onClose={onClose} title={`history: ${label}`} size="xl">
-      {entries.length === 0 ? (
-        <p className="text-xs text-muted-foreground italic">no history yet</p>
-      ) : (
-        <ul className="flex flex-col gap-4">
-          {entries.map((entry) => (
-            <li
-              key={entry.id}
-              className="flex flex-col gap-1.5 rounded-md border border-border-soft bg-subtle p-3"
-            >
-              <div className="flex items-center justify-between gap-2">
-                <span
-                  className={cn(
-                    'rounded-full px-2 py-0.5 text-2xs uppercase tracking-wide',
-                    entry.author === 'user' ? 'bg-accent/10 text-accent' : 'bg-info/10 text-info',
-                  )}
-                >
-                  {entry.author === 'user' ? 'you' : 'agent'}
-                </span>
-                <span className="text-2xs text-muted-foreground">
-                  {formatRelative(entry.createdAt)}
-                </span>
-                <button
-                  type="button"
-                  onClick={() => onRestore(entry)}
-                  title="restore this version"
-                  aria-label="restore"
-                  className="ml-auto flex items-center gap-1 rounded-sm px-1.5 py-0.5 text-2xs text-muted-foreground hover:bg-muted hover:text-foreground"
-                >
-                  <RotateCcw size={10} aria-hidden />
-                  restore
-                </button>
-              </div>
-              {renderAsMarkdown ? (
-                <div className="max-h-40 overflow-hidden text-xs leading-relaxed text-foreground">
-                  <Markdown text={entry.value} className="text-xs" />
-                </div>
-              ) : (
-                <p className="whitespace-pre-wrap text-xs leading-relaxed text-foreground line-clamp-4">
-                  {entry.value}
-                </p>
-              )}
-            </li>
-          ))}
-        </ul>
-      )}
-    </Dialog>
-  );
-}
-
-function formatRelative(iso: string): string {
-  const diff = Date.now() - new Date(iso).getTime();
-  const mins = Math.floor(diff / 60_000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  return `${Math.floor(hrs / 24)}d ago`;
-}
-
-function SummarizerBadge({
-  sessionId,
-  status,
-  lastUpdate,
-  error,
-  totals,
-  canRetry,
-}: {
-  sessionId: SessionId;
-  status: SummarizerStatusKind;
-  lastUpdate: string | null;
-  error: string | null;
-  totals: {
-    readonly inputTokens: number;
-    readonly outputTokens: number;
-    readonly estimatedCostUsd: number;
-    readonly count: number;
-  };
-  canRetry: boolean;
-}) {
-  const retrySummarizer = useAppStore((s) => s.retrySummarizer);
-  const [retrying, setRetrying] = useState(false);
-
-  useEffect(() => {
-    if (status !== 'error') setRetrying(false);
-  }, [status]);
-
-  const costTooltip =
-    totals.count === 0
-      ? 'summarizer has not run yet'
-      : `summary total · ${totals.count} run${totals.count === 1 ? '' : 's'} · ${totals.inputTokens} in / ${totals.outputTokens} out · $${totals.estimatedCostUsd.toFixed(4)}${lastUpdate ? ` · last ${lastUpdate}` : ''}`;
-
-  const costPill = (
-    <span
-      title={costTooltip}
-      className="rounded-full bg-subtle px-2 py-0.5 text-2xs text-muted-foreground"
-    >
-      Σ ${totals.estimatedCostUsd.toFixed(4)}
-    </span>
-  );
-
-  if (status === 'running') {
-    return (
-      <span className="flex items-center gap-1">
-        <Loader2 size={10} aria-hidden className="animate-spin text-info" />
-        {costPill}
-      </span>
-    );
-  }
-
-  if (status === 'error') {
-    const errorTitle = error ? `cannot summarize · ${error}` : 'cannot summarize';
-    return (
-      <span className="flex items-center gap-1">
-        {costPill}
-        <button
-          type="button"
-          onClick={() => {
-            if (!canRetry || retrying) return;
-            setRetrying(true);
-            retrySummarizer(sessionId);
-          }}
-          disabled={!canRetry}
-          title={canRetry ? `${errorTitle}, click to retry` : errorTitle}
-          aria-label={canRetry ? 'retry summarizer' : 'summarizer failed'}
-          className={cn(
-            'inline-flex items-center gap-1 rounded-full bg-danger/10 px-2 py-0.5 text-2xs uppercase tracking-wide text-danger transition-colors',
-            canRetry
-              ? 'hover:bg-danger/15 hover:text-danger-foreground/90'
-              : 'cursor-not-allowed opacity-70',
-          )}
-        >
-          <AlertTriangle size={10} aria-hidden />
-          Cannot summarize
-          <RotateCw size={10} aria-hidden className={cn('shrink-0', retrying && 'animate-spin')} />
-        </button>
-      </span>
-    );
-  }
-
-  return costPill;
 }
